From "Nick Kossifidis" <mickflemm@gmail.com>:
I fixed a bug on ath_info (it didn't update offset after reading tx
power calibration data and only 802.11a values were processed). Also i
added per rate txpower calibration data, but only for 802.11a since b
and g are somewhere else and we need to find the correct offsets.

I believe we have what we need from eeprom to implement txpower
support for RF5112 and 802.11a mode (we need to create the table etc
now). Doing the same for 5111 won't be difficult, the only difference
is that frequency mask is 0x7f instead of 0xff and we have to fix
shifts etc. I am thinking of a function that does this conversion some
way (eg. we provide the value we read and an int[] with masks and it
returns an int[] with the values) to avoid rewriting the same stuff
for 5111 (and possibly for newer chips where power mask changes to
0x7f from 0x3f -0.25 dbm steps).